image-sender/requirements.txt
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
FlightRadarAPI==1.3.25
geographiclib==2.0
geopy==2.4.1
idna==3.6
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.1
pytz==2023.4
pyzmq==25.1.2
requests==2.31.0
snakeviz==2.2.0
tornado==6.4
update-checker==0.18.0
urllib3==2.2.0
websocket-client==1.7.0
